@@ -11,22 +11,22 @@
     "git:push": "(git diff --quiet || (echo 'You have uncommitted changes. Please commit your changes first.' && exit 1)) && git push"
   },
   "dependencies": {
-    "@commitlint/cli": "^17.6.7",
-    "@commitlint/config-conventional": "^17.6.7",
-    "@commitlint/cz-commitlint": "^17.6.7",
+    "@commitlint/cli": "^18.4.3",
+    "@commitlint/config-conventional": "^18.4.3",
+    "@commitlint/cz-commitlint": "^18.4.3",
     "@emailjs/browser": "^3.10.0",
     "@tanstack/react-query": "^4.26.1",
-    "@types/node": "18.14.1",
+    "@types/node": "20.10.5",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
     "eslint": "8.34.0",
-    "eslint-config-next": "13.2.1",
+    "eslint-config-next": "14.0.4",
     "framer-motion": "^10.0.1",
     "husky": "^8.0.3",
     "next": "^14.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "typescript": "4.9.5"
+    "typescript": "5.3.3"
   },
   "devDependencies": {
     "@types/react-transition-group": "^4.4.5",
